[BUGFIX] Context User Aspect must deal with empty fe_user
TSFE->fe_user is set to '' by default, but UserAspect
expects "null".

Extensions like solr set up their own TSFE object which
is fine, however they still access gr_list which is OK,
because it is deprecated. However, a deprecation
message must be thrown, and NOT a strict typing warning.

UserAspect expects either "null" or object of type
AbstractUserAuthentication, but TSFE delivers '' by
default.

Modifying '' would not be an option as this would be
(slightly but still) breaking.

Resolves: #85632
Releases: master
Change-Id: I2dbd46bee1c497f413ede8a1e55334ef954cf723
Reviewed-on: https://review.typo3.org/57674
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>