[FEATURE] Add TCA 'saltedPassword' eval for type=input
The salted passwords extension did register a custom
field evaluation to hash a given password when editing users
in the backend through FormEngine.
This is now turned into a general eval 'saltedPassword'
and handled in the DataHandler directly. Note the existing
eval 'password' is not extended since 'password' eval
changes the display part to show '****', which should be
kept seperated from the salt evaluation.
The "Evaluation" classes are dropped directly without
substitution since they are totally saltedpasswords internal
and don't make sense to be used externally.

Change-Id: I93306eaf18222d14e141a3d612a419a2b7174341
Resolves: #85698
Releases: master
Reviewed-on: https://review.typo3.org/57746
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>