[TASK] Remove dependency between SiteLanguage and Site
The entity of SiteLanguage is re-connected to a
site. However this circular dependency makes it impossible
to serialize or compile any of the site configuration.

As this was introduced at the very early stage of
Site Handling where the API wasn't clear, this can
be removed, as "site" and "language" are both
always available in a PSR-7 request.

Resolves: #85841
Releases: master
Change-Id: Id061def7b0299b9c355ae83e1903cf64b6127149
Reviewed-on: https://review.typo3.org/57893
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>