[TASK] Deprecate PageGenerator class
The static class PageGenerator is a list of
static methods previouly found in "pagegen.php", which
has now been moved within the RequestHandler which
is now fully responsible for turning a request object
with support from TSFE and PageRenderer into a response object.

This removes one further dependency built around a
legacy class and enjoys the concept that RequestHandler+TSFE
now share the responsibility for building $TSFE->content.

This also means that all tests can now be built without
having to use any global variables or some special Singleton
functionality but with a clean structure in mind.

Resolves: #85884
Related: #85822
Releases: master
Change-Id: I298795ea485f3c9850af23c0c5e69a3a6d43df99
Reviewed-on: https://review.typo3.org/57941
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Mathias Brodala <mbrodala@pagemachine.de>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>