[BUGFIX] Ensure cObj only contains valid property objects
If a cObj instance is serialized (which happens very often), all sub-objects are
serialized with it.

This has a bad side-effect with Logging (as we've seen in other places like scheduler tasks)
when a logger with a FileWriter (to a resource pointer) is used, as the __wakeup() basically
destroys everything.

Therefore the logger is instantiated separately in __wakeup() - the same behaviour
as done with TSFE in ContentObjectRenderer.

Additionally, in order to improve memory footprint, the MarkerBasedTemplateService
is only instantiated whenever used (and not serialized).

Resolves: #85886
Releases: master
Change-Id: I3347485c2a6dcb6cf86a4f8182df6aa004582bb5
Reviewed-on: https://review.typo3.org/57944
Reviewed-by: Mathias Brodala <mbrodala@pagemachine.de>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Stephan Großberndt <stephan.grossberndt@typo3.org>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>