[TASK] Deprecate cObject FILE
The cObject FILE has a strange behaviour, and rather became low-level since
Fluid's rise in rendering Frontend.

FILE shows an image if the file is jpg,jpeg,gif,png (but not SVG), and if has a
different file ending, it will check if the file is less than 1MB, it will get its contents
and output that. If the file is 1 byte bigger, nothing is done.

This _magic_ is sketchy, and obsolete, as Fluid offers more flexibility nowadays.

Thus, cObject FILE will be removed from TYPO3 Core in v10.0, however due
to the flexibility of custom cObject, an extension author could just re-implement
the functionality in a custom extension within minutes.

Resolves: #85970
Releases: master
Change-Id: I792ec89486e29051da478c5ec0855d2f9cf913a1
Reviewed-on: https://review.typo3.org/58028
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>