[TASK] Streamline page language linking
When having a Site configuration in order to create a page link for
a specific language new property `language` has to be used instead
of previous `additionalParams=&L=1`. Since it might not be known in
all cases whether the defined TypoScript is used inside or outside
a valid Site configuration, `language` settings now take precedence
and override legacy L-parameter.

Besides that, linking to localized pages was not possible and is
resolved to their language parent pages. The new `language` setting
will be adjusted at the same time as well.

Resolves: #86058
Releases: master
Change-Id: I9531a14f8aa5913a03fdac9fcbdaced57312c2af
Reviewed-on: https://review.typo3.org/58101
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Frans Saris <franssaris@gmail.com>
Tested-by: Frans Saris <franssaris@gmail.com>