[BUGFIX] Allow page links to different languages within a site
Due to the refactoring process within links to a site, it is still
necessary to link to a translated page with a given "&L=" parameter.
This parameter is resolved, and a link to a page translation
is generated.

Fixes menu generation and typolink with the generated page title.

Resolves: #86067
Releases: master
Change-Id: I3e1208a2cdb438c68d4ed3dac1d0274ce07395dc
Reviewed-on: https://review.typo3.org/58108
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Susanne Moog <susanne.moog@typo3.org>
Tested-by: Susanne Moog <susanne.moog@typo3.org>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Tested-by: Oliver Hader <oliver.hader@typo3.org>