[BUGFIX] Ensure that slugs are always calculated in DataHandler
When NOT using FormEngine to create a new page (e.g. in the pagetree
or via the "Create multiple pages" wizard), so just using DataHandler
magic is uttermost critical that a slug is always created, to ensure
the "uniqueInSite" functionality, and to access a page, even though
the page was created after the upgrade wizard has run.

Resolves: #86050
Resolves: #85937
Releases: master
Change-Id: I9eff3385c369a04a6f5a33d0b840b6a2b698891c
Reviewed-on: https://review.typo3.org/58098
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Susanne Moog <susanne.moog@typo3.org>
Tested-by: Susanne Moog <susanne.moog@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>