[BUGFIX] Use correct language in l18n_cfg checks
The check for pages.l18n_cfg is wrong now, as this is
always done against pages.sys_language_uid=0 records (as
"resolvePage" is called right before).

So, these if-statements have to go further south.

On top, the getPageOverlay call needs to be done - on top.

resolvePage -> get default language of page
getPageOverlay -> put the wanted translation on top

A fix for pages.alias has to be in place, as DataHandler
cannot handle pages.alias yet.

Resolves: #86242
Releases: master
Change-Id: Ief99e5f934f6e9d31973b9543cb9a6e599d2d33c
Reviewed-on: https://review.typo3.org/58267
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Tested-by: Oliver Hader <oliver.hader@typo3.org>