[BUGFIX] Allow linking to a targetSite without currentSite
In case no current site exists in a context (for example currently
eID or other scripting contexts), the pageLinkBuilder threw an error
when linking to a page with a site configuration. The necessity for
having a current site config has been removed, the pageLinkBuilder
can now be used without a current site object and will then always
generate absolute URLs to the target site.

Resolves: #86384
Releases: master
Change-Id: Id8c984c4ac3837d7b4da37a99d43410c6db34187
Reviewed-on: https://review.typo3.org/58397
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Henning Liebe <h.liebe@neusta.de>
Tested-by: Henning Liebe <h.liebe@neusta.de>
Reviewed-by: Frank Naegler <frank.naegler@typo3.org>
Tested-by: Frank Naegler <frank.naegler@typo3.org>