[TASK] Add facades for symfony/routing components
We heavily rely on Symfony Routing for TYPO3 routing,
however we want to encapsulate this as much as possible
to allow us to use a different routing system if a better one
comes up.

In order to proceed with Route Enhancers we introduce
our own objects to extend them, and to typehint against
these objects in the next patches.

Some minor code preparations have been made as well.

Resolves: #86393
Releases: master
Change-Id: Id50186aa1a5ad57e097e6ff85e15d0f925bd63ef
Reviewed-on: https://review.typo3.org/58401
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Tested-by: Oliver Hader <oliver.hader@typo3.org>