[TASK] phpDoc: replace @access private with @internal
The phpDoc comment "@internal" is used in most places
during the development of TYPO3 v7, TYPO3 v8 and TYPO3 v9
and denotes that this specific method, class or property is not
part of the Public API, and does not even need any notice if it
is removed.

However, we still have/had various places in TYPO3 core
where "@access private" was used. This is now migrated to
"@internal". "@access public" is not needed and is removed.

Resolves: #86446
Releases: master
Change-Id: Iaef206b04aee13c280d62b107dd7eac04a38c700
Reviewed-on: https://review.typo3.org/58462
Reviewed-by: Susanne Moog <susanne.moog@typo3.org>
Tested-by: Susanne Moog <susanne.moog@typo3.org>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>