[BUGFIX] Only fetch default language page for tree lists
Due to the removal of pages_language_overlay the queries
for fetching trees have to work as before - mainly because the
concept of the "pid" (= storagePid) works the same - records on a
page have the pid of the default language, not of any translation.

Therefore, the used queries have to be built to check for
sys_language_uid=0.

Resolves: #87020
Related: #86961
Releases: master
Change-Id: Ie7740b6a32618a213d569362a58b9f8e5ed7824d
Reviewed-on: https://review.typo3.org/58967
Reviewed-by: Richard Haeser <richard@maxserv.com>
Tested-by: Richard Haeser <richard@maxserv.com>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>