[BUGFIX] Enable preview of hidden page translations
TSFE->determineIdIsHiddenPage() now takes the current site language into
account when fetching the page record. When rendering a page translation
the method does no longer reflect the hidden status of the default
language but the status of the translated page instead.

This enables previewing a hidden page translation even if the language
parent is active.

Resolves: #85471
Releases: master, 9.5
Change-Id: I1b0ead8eaf4c707583e24babbbca49870a31cdaa
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/60364
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Daniel Goerz <daniel.goerz@posteo.de>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Daniel Goerz <daniel.goerz@posteo.de>
Reviewed-by: Benni Mack <benni@typo3.org>