[FEATURE] Introduce PreviewAspect for handling the preview flag
The public internal property of TSFE "fePreview" has been deprecated
in favor of a dedicated aspect handling the frontend preview flag.

As a first step, the frontend preview aspect only contains a boolean
flag holding the current preview state - but it might be extended with
additional information on the preview mode at a later point.

This is also a first step to remove the TSFE dependencies from the
AdminPanel initialisation process.

Resolves: #88791
Releases: master
Change-Id: I2994d69d824b1b107a3a878b87c9dd2f7daf3ed2
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/61311
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Frank Naegler <frank.naegler@typo3.org>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>