[BUGFIX] Support multiple order by fields in PageRepository
The method PageRepository::getSubpagesForPages allows an arbitrary
number of orderBy fields, though the way it is passed to the
query builder, only the last field is used.

This patch simply changes the DBAL call to ->addOrderBy to ensure
all set order by fields are used, in the order they are defined.

Resolves: #88876
Releases: master, 9.5
Change-Id: I97ad77490ec42cec5e46ec954bf4e2ba694d9eec
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/61371
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Frank Naegler <frank.naegler@typo3.org>
Reviewed-by: Frank Naegler <frank.naegler@typo3.org>