[BUGFIX] Fix frontend error handling when root page is protected
If the root page is accessed directly, we add a check whether
the page could be fetched with ignored group permissions.
If that is the case, an access denied is triggered.

If the root page also has the extendToSubPages flag set,
we must not remove the page from the root line, which would
trigger a service unavailable handling, but instead
keep the root line and trigger a page not found.

Resolves: #89006
Releases: master, 9.5
Change-Id: I69fa1c725c16d69c37caca5e3013e0c010854b5f
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/61596
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>