[BUGFIX] Make the $target parameter for addSecurityRelValues nullable
In the typoLink() function of ContentObjectRenderer, not all links will
have a 'target' value. - If the protocol is not natively supported by
TYPO3, as in "tel:", only a 'url' parameter will be resolved.
For that reason, make the $target parameter of addSecurityRelValues()
nullable to avoid frontend errors when using custom link protocols.

Resolves: #89338
Releases: master
Change-Id: I6bca872e38e2f5618a5c06a5ebd3476beb2e60a0
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/61879
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Susanne Moog <look@susi.dev>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Daniel Goerz <daniel.goerz@posteo.de>
Reviewed-by: Susanne Moog <look@susi.dev>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>