[BUGFIX] Ignore includeNotInMenu for the section menu of the current page
The option includeNotMenu is set to 1 if the CType menu_section is used
to render a section menu for the current page even if the page is hidden
in menus (pages.nav_hide).

This patch adds also the necessary stdWrap functionality to the abstract
class for hierarchical menus.

Releases: master, 9.5
Resolves: #91044
Change-Id: I563390b4af36fb26e6a02df237563bdc48293048
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/64183
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Benni Mack <benni@typo3.org>