[BUGFIX] Use correct slug for access restricted translated pages
Access restricted, translated pages currently always have the slug from the
default language instead of their translated slug in the frontend.
To generate correct urls, while using the "linkAccessRestrictedPages"
option, the $disableGroupAccessCheck parameter needs to also take into
account for
 * PageRepository::getPageOverlay()
when setting the option in PageRepository::getPage().

This "hack" is currently similar to what HMENU is doing, however
this public property should not be used with the Context API instead.

This change however needs more refactoring on the Context API,
which is why this solution is chosen for the time being (and also
for v9 backport).

Resolves: #90842
Resolves: #87969
Resolves: #91185
Releases: master, 9.5
Change-Id: I99a34ca7fceacba7218c6b7132781805a6b59ac9
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/63963
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Jonas Eberle <flightvision@googlemail.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Jonas Eberle <flightvision@googlemail.com>
Reviewed-by: Helmut Hummel <typo3@helhum.io>
Reviewed-by: Benni Mack <benni@typo3.org>