[BUGIFX] Store state of PageTitleProviderManager in cache
To allow uncached plugins to provide a different page title,
page title providers must be also be called after uncached plugins
are rendered. To not loose the state of page title providers
that are called for cached entires, the state of page title
providers is stored in cache as well and provided for the
second title generation process after uncached objects are rendered.

Resolves: #91233
Releases: master, 10.4, 9.5

Change-Id: I277551d9a58781c0d130c27b346bcbbc209266fd
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/64972
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Daniel Goerz <daniel.goerz@posteo.de>
Reviewed-by: Daniel Goerz <daniel.goerz@posteo.de>